refactor(clickhouse): isolate adapter utilities
